# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/supplements": {
      "get": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supplements"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSupplementRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSupplementRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSupplementRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/{id}": {
      "put": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSupplementRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSupplementRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSupplementRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/supplements/{id}/active": {
      "patch": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetSupplementActiveRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetSupplementActiveRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetSupplementActiveRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/{id}/slots": {
      "post": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/{id}/slots/{slotId}": {
      "put": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "slotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SupplementSlotRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "slotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/checklist": {
      "get": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "timing",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SupplementTimingEnum"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/intakes": {
      "get": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementIntakeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementIntakeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplementIntakeDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Supplements"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ToggleIntakeRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ToggleIntakeRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ToggleIntakeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supplements"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LogAdHocIntakeRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LogAdHocIntakeRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LogAdHocIntakeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/supplements/intakes/{intakeId}": {
      "delete": {
        "tags": [
          "Supplements"
        ],
        "parameters": [
          {
            "name": "intakeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplementChecklistDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSupplementRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplementUnitEnum"
              }
            ]
          },
          "defaultAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LogAdHocIntakeRequest": {
        "type": "object",
        "properties": {
          "supplementId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "workoutSessionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SetSupplementActiveRequest": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SupplementChecklistDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplementChecklistItemDto"
            },
            "nullable": true
          },
          "adHoc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplementIntakeDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SupplementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplementUnitEnum"
              }
            ]
          },
          "defaultAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplementSlotDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplementIntakeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "supplementId": {
            "type": "integer",
            "format": "int32"
          },
          "supplementName": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplementUnitEnum"
              }
            ]
          },
          "scheduleSlotId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "takenOn": {
            "type": "string",
            "format": "date"
          },
          "takenAt": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "workoutSessionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SupplementSlotRequest": {
        "type": "object",
        "properties": {
          "timing": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplementTimingEnum"
              }
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "timeOfDay": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "offsetMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplementTimingEnum": {
        "enum": [
          "Morning",
          "Midday",
          "Afternoon",
          "Evening",
          "Night",
          "PreWorkout",
          "PostWorkout",
          "WithMeal",
          "Custom"
        ],
        "type": "string"
      },
      "SupplementUnitEnum": {
        "enum": [
          "Piece",
          "Capsule",
          "Tablet",
          "Gram",
          "Milligram",
          "Milliliter",
          "Scoop",
          "Drop",
          "InternationalUnit"
        ],
        "type": "string"
      },
      "ToggleIntakeRequest": {
        "type": "object",
        "properties": {
          "supplementId": {
            "type": "integer",
            "format": "int32"
          },
          "slotId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "taken": {
            "type": "boolean"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "workoutSessionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateSupplementRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplementUnitEnum"
              }
            ]
          },
          "defaultAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter ONLY token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "x-time-zone": {
        "type": "apiKey",
        "description": "Specify the user's timezone in IANA format (e.g., 'America/New_York')",
        "name": "x-time-zone",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    },
    {
      "x-time-zone": [ ]
    }
  ]
}